server/requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
blinker==1.9.0
boto3==1.37.17
botocore==1.37.17
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
ollama==0.4.7
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydicom==3.0.1
PyJWT==2.10.1
PyMuPDF==1.25.3
PyMySQL==1.1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.4
safetensors==0.5.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
